EnemyRefactoring

+++ GamaJama/Assets/Code/Scripts/Interfaces/IPlayerDamager.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IPlayerDamager
 {
-   public float GetDamage();
+   public int GetDamage();
 }
